% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silnik_psychometria_analizy.R
\name{excel_file_name}
\alias{excel_file_name}
\title{Nazwa pliku Excel z wynikami}
\usage{
excel_file_name(data_path, time = Sys.time())
}
\arguments{
\item{data_path}{Sciezka do pliku danych wejsciowych.}

\item{time}{Czas uzyty w znaczniku (domyslnie biezacy).}
}
\value{
Nazwa pliku (bez katalogu).
}
\description{
Tworzy nazwe pliku \verb{<plik_danych>_results_<RRRR-MM-DD_GGMMSS>.xlsx}.
Znacznik czasu sprawia, ze kolejne rendery (np. wersje EN i PL raportu)
nie nadpisuja wczesniejszych plikow.
}
\examples{
excel_file_name("dane/math_data.csv", as.POSIXct("2026-10-05 10:28:19"))

}
