% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fmt_table}
\alias{fmt_table}
\title{Formatowanie tabeli do raportu}
\usage{
fmt_table(df, ...)
}
\arguments{
\item{df}{Ramka danych.}

\item{...}{Dodatkowe argumenty przekazywane do \code{\link[kableExtra:kbl]{kableExtra::kbl()}}.}
}
\value{
Obiekt klasy \code{kableExtra}.
}
\description{
Tlumaczy nazwy kolumn (tokeny \verb{col.*}, zob. \code{\link[=t_kolumny]{t_kolumny()}})
i formatuje ramke danych jako tabele kableExtra w stylu
striped/hover/condensed.
}
\examples{
fmt_table(data.frame(Item = "i1", Trudnosc_p = 0.5))

}
