% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silnik_psychometria_wczytanie_walidacja.R
\name{identify_item_columns}
\alias{identify_item_columns}
\title{Identyfikacja kolumn itemow}
\usage{
identify_item_columns(raw_data, item_prefix, exclude_items = NULL)
}
\arguments{
\item{raw_data}{Ramka danych}

\item{item_prefix}{Prefiks nazw kolumn z itemami.}

\item{exclude_items}{Opcjonalny wektor nazw itemow do wykluczenia.}
}
\value{
Wektor nazw kolumn z itemami.
}
\description{
Funkcja wyszukuje kolumny rozpoczynajace sie od wskazanego prefiksu.
Jezeli itemy maja numeryczne koncowki, porzadkuje je wedlug numerow.
}
\details{
Funkcja jest czysta: nie zapisuje plikow, nie wypisuje komunikatow
i nie modyfikuje przekazanej ramki danych.
}
