% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silnik_psychometria_analizy.R
\name{interpret_alpha}
\alias{interpret_alpha}
\title{Interpretacja wspolczynnika alfa Cronbacha}
\usage{
interpret_alpha(alpha_val)
}
\arguments{
\item{alpha_val}{Jedna wartosc liczbowa reprezentujaca alfa Cronbacha.}
}
\value{
Jednoelementowy wektor tekstowy z interpretacja rzetelnosci.
}
\description{
Przypisuje tekstowa interpretacje do wartosci wspolczynnika alfa Cronbacha zgodnie z progami stosowanymi w raporcie psychometrycznym.
}
\examples{
interpret_alpha(0.82)

}
