% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silnik_psychometria_analizy.R
\name{interpret_omega}
\alias{interpret_omega}
\title{Interpretacja wspolczynnika omega McDonalda}
\usage{
interpret_omega(omega_val)
}
\arguments{
\item{omega_val}{Jedna wartosc liczbowa reprezentujaca omega McDonalda.}
}
\value{
Jednoelementowy wektor tekstowy z interpretacja rzetelnosci.
}
\description{
Przypisuje tekstowa interpretacje do wartosci wspolczynnika omega McDonalda zgodnie z progami stosowanymi w raporcie psychometrycznym.
}
\examples{
interpret_omega(0.76)

}
