% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{math_data}
\alias{math_data}
\title{Przykładowe dane z testu matematycznego}
\format{
Ramka danych o 2448 wierszach i 52 kolumnach:
\describe{
\item{id_szkoly}{Identyfikator szkoły}
\item{id_ucznia}{Identyfikator ucznia}
\item{pomiar}{Numer pomiaru}
\item{grupa}{Grupa}
\item{mat_1, mat_2, ..., mat_48}{Odpowiedzi na zadania matematyczne
(0 = błędna, 1 = poprawna, NA = brak odpowiedzi)}
}
}
\usage{
math_data
}
\description{
Zbiór danych zawierający odpowiedzi uczniów na zadania z testu
matematycznego. Dane zawierają odpowiedzi binarne (0/1) na 48 zadań
oraz informacje identyfikacyjne.
}
\keyword{datasets}
