% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silnik_psychometria_analizy.R
\name{safe_sheet_name}
\alias{safe_sheet_name}
\title{Przygotowanie bezpiecznej nazwy arkusza Excel}
\usage{
safe_sheet_name(x)
}
\arguments{
\item{x}{Jednoelementowy wektor tekstowy z proponowana nazwa arkusza.}
}
\value{
Jednoelementowy wektor tekstowy zawierajacy bezpieczna nazwe arkusza.
}
\description{
Oczyszcza proponowana nazwe arkusza z niedozwolonych znakow i przycina ja do 31 znakow, czyli limitu stosowanego przez Excel. Funkcja jest uzywana podczas eksportu wynikow do pliku XLSX.
}
\examples{
safe_sheet_name("Wyniki Analizy CTT (Wersja A)!")

}
