% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{show_table}
\alias{show_table}
\title{Wyswietlenie tabeli w raporcie}
\usage{
show_table(df, ...)
}
\arguments{
\item{df}{Ramka danych.}

\item{...}{Dodatkowe argumenty przekazywane do \code{\link[=fmt_table]{fmt_table()}}.}
}
\value{
Niewidocznie \code{df}. Funkcja jest wywolywana dla efektu ubocznego.
}
\description{
Formatuje ramke danych przez \code{\link[=fmt_table]{fmt_table()}} i wypisuje kod
HTML tabeli (do chunkow z \code{results = "asis"}). Nie korzysta z metody
\code{print()} kableExtra, ktora w sesji interaktywnej (np. render z konsoli
RStudio) wysyla tabele do panelu Viewer zamiast do raportu.
}
\examples{
show_table(data.frame(Item = "i1", Trudnosc_p = 0.5))

}
