% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlumaczenia.R
\name{t}
\alias{t}
\title{Tlumaczenie tokena}
\usage{
t(token)
}
\arguments{
\item{token}{Wektor tekstowy z tokenami.}
}
\value{
Wektor tekstowy z tlumaczeniami, tej samej dlugosci co \code{token}.
}
\description{
Zwraca tlumaczenie tokena w jezyku wybranym przez
\code{\link[=wczytaj_tlumaczenia]{wczytaj_tlumaczenia()}}. Jezeli tlumaczenia nie zostaly jeszcze wczytane,
wczytywany jest jezyk domyslny (angielski). Dla tokenow nieobecnych w pliku tlumaczen
zwracany jest sam token, dzieki czemu braki sa widoczne w raporcie.

Funkcja przeslania \code{base::t()}. Dla argumentow nietekstowych oraz macierzy
przekazuje wywolanie do \code{base::t()}, wiec transpozycja dziala jak dotad.
}
\examples{
t("common.whole_test")

}
