% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlumaczenia.R
\name{t_kolumny}
\alias{t_kolumny}
\title{Tlumaczenie nazw kolumn ramki danych}
\usage{
t_kolumny(df)
}
\arguments{
\item{df}{Ramka danych.}
}
\value{
Ramka danych z przetlumaczonymi nazwami kolumn.
}
\description{
Zamienia nazwy kolumn na tlumaczenia tokenow
\verb{col.<nazwa_kolumny>}. Kolumny bez tokena zachowuja oryginalna nazwe.
Funkcja sluzy wylacznie do wyswietlania tabel - w kodzie pakietu nazwy
kolumn pozostaja niezmienione.
}
\examples{
t_kolumny(data.frame(Item = "i1", Trudnosc_p = 0.5))

}
