% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlumaczenia.R
\name{t_logiczne}
\alias{t_logiczne}
\title{Tlumaczenie wartosci logicznych w ramce danych}
\usage{
t_logiczne(df)
}
\arguments{
\item{df}{Ramka danych.}
}
\value{
Ramka danych, w ktorej kolumny logiczne sa tekstowe.
}
\description{
Zamienia kolumny logiczne na tekst z tokenow \code{common.yes}
i \code{common.no} (np. Yes/No, Tak/Nie). Braki (\code{NA}) zostaja brakami.
Funkcja sluzy wylacznie do wyswietlania i eksportu tabel - w kodzie
pakietu kolumny pozostaja logiczne. Bez tej zamiany Excel wyswietla
wartosci logiczne w jezyku swojej instalacji (np. PRAWDA/FALSZ),
niezaleznie od jezyka raportu.
}
\examples{
t_logiczne(data.frame(Item = c("i1", "i2"), DIF_signal = c(TRUE, FALSE)))

}
