% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silnik_psychometria_wczytanie_walidacja.R
\name{validate_items_data}
\alias{validate_items_data}
\title{Walidacja itemow testowych}
\usage{
validate_items_data(raw_data, item_cols)
}
\arguments{
\item{raw_data}{Ramka danych}

\item{item_cols}{Wektor nazw kolumn z itemami}
}
\value{
Lista z oczyszczonymi danymi itemowymi, diagnostykami walidacyjnymi
oraz klasyfikacja typu itemow (\code{item_type}, \code{item_max_scores},
\code{n_categories}).
}
\description{
Sprawdza typ danych, klasyfikuje itemy jako binarne lub
politomiczne i weryfikuje wariancje. Itemy z prawidlowym kodowaniem
(kolejne liczby calkowite od 0 do max) sa zachowywane niezaleznie od
liczby kategorii. Itemy z niecalkowitymi, ujemnymi lub niespojnymi
wartosciami sa wykluczane.
}
\details{
Funkcja jest czysta wzgledem srodowiska zewnetrznego: nie zapisuje
plikow, nie wypisuje komunikatow i nie modyfikuje przekazanej ramki danych.
Zwraca oczyszczone dane itemowe oraz diagnostyki walidacyjne.
}
