% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlumaczenia.R
\name{wczytaj_tlumaczenia}
\alias{wczytaj_tlumaczenia}
\title{Wczytanie tlumaczen raportu}
\usage{
wczytaj_tlumaczenia(jezyk = "en", sciezka = NULL)
}
\arguments{
\item{jezyk}{Kod jezyka odpowiadajacy nazwie kolumny w pliku tlumaczen,
np. \code{"en"} (domyslnie) lub \code{"pl"}.}

\item{sciezka}{Opcjonalna sciezka do pliku CSV z tlumaczeniami. Domyslnie
plik \code{tlumaczenia.csv} z katalogu \code{reports} zainstalowanego pakietu.}
}
\value{
Niewidocznie: nazwany wektor tekstowy z tlumaczeniami.
}
\description{
Wczytuje tlumaczenia z pliku CSV (separator \verb{;}, kolumna
\code{token} oraz po jednej kolumnie na kazdy jezyk, np. \code{en}, \code{pl}), buduje
wektor tlumaczen w wybranym jezyku, nazwany tokenami, i zapisuje go jako
opcje konfiguracyjna \code{tlumaczenia}. Z opcji korzysta funkcja \code{\link[=t]{t()}}.
}
\examples{
wczytaj_tlumaczenia("pl")
t("common.whole_test")
wczytaj_tlumaczenia()

}
