% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silnik_psychometria_analizy.R
\name{z_score}
\alias{z_score}
\title{Standaryzacja zmiennej do wyniku z}
\usage{
z_score(x)
}
\arguments{
\item{x}{Wektor liczbowy przeznaczony do standaryzacji.}
}
\value{
Wektor liczbowy tej samej dlugosci co \code{x}, zawierajacy wartosci z.
}
\description{
Przelicza wektor liczbowy na wynik standaryzowany z, odejmujac srednia i dzielac przez odchylenie standardowe. Funkcja obsluguje braki danych oraz zwraca brak danych, gdy odchylenie standardowe jest zerowe lub niemozliwe do wyznaczenia.
}
\examples{
z_score(c(1, 2, 3, 4, 5))

}
