% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silnik_psychometria_analizy.R
\name{make_empirical_icc_plot}
\alias{make_empirical_icc_plot}
\title{Przygotowanie wykresu empirycznego dopasowania ICC}
\usage{
make_empirical_icc_plot(
  model,
  data_items,
  theta_vals,
  label = "Caly test",
  model_name = "IRT"
)
}
\arguments{
\item{model}{Dopasowany model IRT z pakietu \code{mirt}.}

\item{data_items}{Ramka danych lub macierz odpowiedzi itemowych.}

\item{theta_vals}{Wektor liczbowy z oszacowaniami theta dla osob.}

\item{label}{Etykieta analizy uzywana w tytule wykresu.}

\item{model_name}{Nazwa modelu IRT uzywana w tytule wykresu.}
}
\value{
Obiekt klasy \code{ggplot} albo \code{NULL}, gdy nie da sie utworzyc grup theta.
}
\description{
Porownuje przewidywane krzywe charakterystyczne itemow z empirycznymi srednimi odpowiedziami w grupach theta i zwraca jeden obiekt ggplot z panelami dla itemow.
}
\examples{
# make_empirical_icc_plot(model, data_items, theta_vals)

}
