% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silnik_psychometria_analizy.R
\name{make_irt_plots}
\alias{make_irt_plots}
\title{Przygotowanie wykresow dla wybranego modelu IRT}
\usage{
make_irt_plots(
  preferred_model,
  data_items,
  theta_vals,
  label,
  preferred_name,
  show_empirical_icc = TRUE
)
}
\arguments{
\item{preferred_model}{Dopasowany model IRT wybrany do interpretacji.}

\item{data_items}{Ramka danych lub macierz itemow uzytych w modelu.}

\item{theta_vals}{Wektor liczbowy z oszacowaniami theta.}

\item{label}{Etykieta analizy uzywana w tytulach wykresow.}

\item{preferred_name}{Nazwa wybranego modelu IRT.}

\item{show_empirical_icc}{Wartosc logiczna okreslajaca, czy przygotowac wykres empirycznego dopasowania ICC.}
}
\value{
Lista obiektow wykresow.
}
\description{
Tworzy liste obiektow wykresow dla wybranego modelu IRT, obejmujaca funkcje informacyjna testu, krzywe charakterystyczne itemow, opcjonalne dopasowanie empiryczne ICC oraz histogram theta.
}
\examples{
# make_irt_plots(model, data_items, theta_vals, "Caly test", "2PL")

}
