% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silnik_psychometria_analizy.R
\name{make_params_table}
\alias{make_params_table}
\title{Przygotowanie tabeli parametrow IRT}
\usage{
make_params_table(model, model_name = NA_character_)
}
\arguments{
\item{model}{Dopasowany model IRT z pakietu \code{mirt}.}

\item{model_name}{Nazwa modelu zapisywana w tabeli, np. \code{"1PL"}, \code{"2PL"} lub \code{"3PL"}.}
}
\value{
Ramka danych z parametrami itemow IRT.
}
\description{
Pobiera parametry itemow z dopasowanego modelu IRT i porzadkuje je w ramce danych zawierajacej dyskryminacje, trudnosc, zgadywanie oraz interpretacje dyskryminacji.
}
\examples{
# model <- mirt::mirt(data_items, 1, itemtype = "2PL")
# make_params_table(model, "2PL")

}
