% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_report.R
\name{render_report}
\alias{render_report}
\title{Generuje raport psychometryczny}
\usage{
render_report(
  output_path,
  data_path,
  item_prefix,
  group_var,
  id_var,
  dif_group_var
)
}
\arguments{
\item{output_path}{ścieżka pod którą zapisany zostanie wygenerowany raport}

\item{data_path}{ścieżka do pliku z danymi TODO-opisać format}

\item{item_prefix}{TODO-opisać}

\item{group_var}{nazwa zmiennej dzielącej analizowane obserwacje na grupy}

\item{id_var}{nazwa zmiennej unikalnie identyfikującej obserwację w zbiorze danych}

\item{dif_group_var}{TODO-opisać}
}
\description{
Generuje raport psychometryczny ze wskazanych danych
}
\examples{
\dontrun{
data_path <- system.file("extdata", "math_data.csv", package = "aazbie")
render_report("raport.html", data_path, "mat_", "grupa", "id_ucznia", "grupa")
}

}
