% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silnik_psychometria_analizy.R
\name{run_ctt_for_items}
\alias{run_ctt_for_items}
\title{Analiza klasycznej teorii testów (CTT) dla zestawu itemow}
\usage{
run_ctt_for_items(
  data_items,
  label = t("common.whole_test"),
  correlation_threshold_negative = 0,
  item_max_scores = NULL
)
}
\arguments{
\item{data_items}{Ramka danych lub macierz z odpowiedziami na itemy. Kolumny odpowiadaja itemom, a wiersze osobom.}

\item{label}{Etykieta analizy zapisywana w wyniku i uzywana przy tworzeniu obiektow wykresow.}

\item{correlation_threshold_negative}{Prog korelacji \code{r_cor}, ponizej ktorego itemy sa zwracane jako slabe w polu \code{weak_items}.}

\item{item_max_scores}{Opcjonalny named vector z maksymalnym wynikiem per item. Gdy NULL, obliczany z danych. Uzywany do normalizacji trudnosci dla itemow politomicznych.}
}
\value{
Lista zawierajaca status, statystyki rzetelnosci, statystyki itemow, podsumowania, wykresy i oczyszczone dane itemowe.
}
\description{
Oblicza podstawowe wskazniki CTT dla przekazanych itemow, w tym rzetelnosc, statystyki itemow, klasyfikacje jakosci itemow oraz obiekt wykresu mapy itemow.
}
\examples{
data_items <- data.frame(i1 = c(0, 1, 1), i2 = c(1, 1, 0), i3 = c(0, 0, 1))
run_ctt_for_items(data_items, "Przyklad")

}
