% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silnik_psychometria_analizy.R
\name{run_dif_analysis}
\alias{run_dif_analysis}
\title{Analiza DIF dla wszystkich par grup}
\usage{
run_dif_analysis(
  raw_data,
  items_data,
  irt_results,
  dif_group_var = NULL,
  group_var = NULL,
  has_groups = FALSE,
  has_versions = FALSE,
  detected_version_col = "detected_version"
)
}
\arguments{
\item{raw_data}{Ramka danych z danymi zrodlowymi, w tym zmienna grupujaca.}

\item{items_data}{Ramka danych lub macierz z odpowiedziami itemowymi.}

\item{irt_results}{Lista wynikow IRT, zwykle zawierajaca element \code{all} lub elementy dla wersji testu.}

\item{dif_group_var}{Nazwa zmiennej grupujacej uzywanej w analizie DIF.}

\item{group_var}{Alternatywna nazwa zmiennej grupujacej, uzywana gdy \code{has_groups = TRUE}.}

\item{has_groups}{Wartosc logiczna informujaca, czy dane maja zdefiniowana zmienna grupujaca.}

\item{has_versions}{Wartosc logiczna informujaca, czy analize wykonac osobno dla wykrytych wersji testu.}

\item{detected_version_col}{Nazwa kolumny w \code{raw_data} zawierajacej wykryta wersje testu.}
}
\value{
Lista zawierajaca status, nazwe zmiennej grupujacej, poziomy grup oraz wyniki analiz DIF dla par grup.
}
\description{
Uruchamia analize DIF dla wszystkich par wartosci zmiennej grupujacej, osobno dla wersji testu lub dla calego zestawu danych, zaleznie od ustawien.
}
\examples{
# run_dif_analysis(raw_data, items_data, list(all = irt), dif_group_var = "plec")

}
