% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silnik_psychometria_analizy.R
\name{run_irt_for_items}
\alias{run_irt_for_items}
\title{Analiza IRT dla zestawu itemow}
\usage{
run_irt_for_items(
  data_items,
  label = "Caly test",
  show_plots = TRUE,
  show_empirical_icc = TRUE
)
}
\arguments{
\item{data_items}{Ramka danych lub macierz z odpowiedziami na itemy.}

\item{label}{Etykieta analizy zapisywana w wyniku i uzywana przy tworzeniu wykresow.}

\item{show_plots}{Wartosc logiczna okreslajaca, czy przygotowac obiekty wykresow IRT.}

\item{show_empirical_icc}{Wartosc logiczna okreslajaca, czy przygotowac wykres empirycznego dopasowania ICC.}
}
\value{
Lista zawierajaca status, modele IRT, porownania modeli, parametry itemow, wyniki theta, wykresy i dane uzyte w analizie.
}
\description{
Dopasowuje modele IRT 1PL, 2PL oraz opcjonalnie 3PL, porownuje ich dopasowanie, wybiera model preferowany na podstawie testow LRT i zwraca parametry itemow, wyniki theta oraz obiekty wykresow.
}
\examples{
# run_irt_for_items(data_items, "Caly test", show_plots = FALSE)

}
