% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silnik_psychometria_analizy.R
\name{run_item_fit}
\alias{run_item_fit}
\title{Analiza dopasowania itemow w modelu IRT}
\usage{
run_item_fit(
  irt_result,
  label = "Caly test",
  run_pvq1 = FALSE,
  pvq1_n_max = 500,
  pvq1_items_max = 30,
  alpha = 0.05
)
}
\arguments{
\item{irt_result}{Lista wynikowa zwrocona przez \code{run_irt_for_items()}.}

\item{label}{Etykieta analizy zapisywana w wyniku i uzywana w tytule wykresu.}

\item{run_pvq1}{Wartosc logiczna okreslajaca, czy probowac obliczyc statystyke PV-Q1*.}

\item{pvq1_n_max}{Maksymalna liczba obserwacji, dla ktorej obliczana jest PV-Q1*.}

\item{pvq1_items_max}{Maksymalna liczba itemow, dla ktorej obliczana jest PV-Q1*.}

\item{alpha}{Poziom istotnosci uzywany do flagowania niedopasowania S-X2.}
}
\value{
Lista zawierajaca status, surowe wyniki dopasowania, ramki danych ze
statystykami, status PV-Q1* oraz obiekt wykresu.
}
\description{
Oblicza statystyki dopasowania itemow dla wybranego modelu IRT.
Funkcja obsluguje modele binarne oraz modele partial-credit dla itemow
punktowanych politomicznie. Zwraca statystyke S-X2, opcjonalnie infit/outfit
oraz wykres diagnostyczny, bez wypisywania wynikow do konsoli.
}
\examples{
# irt <- run_irt_for_items(data_items, show_plots = FALSE)
# run_item_fit(irt)

}
