% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silnik_psychometria_analizy.R
\name{sequential_elimination}
\alias{sequential_elimination}
\title{Sekwencyjna eliminacja itemow wedlug korelacji item-test}
\usage{
sequential_elimination(
  data_items,
  label = "Caly test",
  thresholds = c(0, 0.1, 0.15)
)
}
\arguments{
\item{data_items}{Ramka danych lub macierz z odpowiedziami na itemy.}

\item{label}{Etykieta analizy zapisywana w wyniku.}

\item{thresholds}{Wektor liczbowy z kolejnymi progami korelacji \code{r_cor} uzywanymi do usuwania itemow.}
}
\value{
Lista zawierajaca status, tabele krokow eliminacji, nazwy pozostalych i usunietych itemow oraz finalne dane.
}
\description{
Wykonuje kolejne kroki eliminacji itemow na podstawie progow skorygowanej korelacji item-test i zwraca historie zmian wraz z finalnym zestawem itemow.
}
\examples{
data_items <- data.frame(i1 = c(0, 1, 1), i2 = c(1, 1, 0), i3 = c(0, 0, 1))
sequential_elimination(data_items)

}
