% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silnik_psychometria_wczytanie_walidacja.R
\name{validate_items_data}
\alias{validate_items_data}
\title{Walidacja itemow testowych}
\usage{
validate_items_data(raw_data, item_cols)
}
\arguments{
\item{raw_data}{Ramka danych}

\item{item_cols}{Wektor nazw kolumn z itemami}
}
\value{
Lista z informacjami o problemach walidacyjnych
}
\description{
Sprawdza typ danych, binarnosc itemow oraz wariancje
}
\details{
Funkcja jest czysta wzgledem srodowiska zewnetrznego: nie zapisuje
plikow, nie wypisuje komunikatow i nie modyfikuje przekazanej ramki danych.
Zwraca oczyszczone dane itemowe oraz diagnostyki walidacyjne.
}
